+++ src/shoobx/immutable/pjpersist.py
@@ -96,6 +96,8 @@ class Immutable(revisioned.RevisionedImmutable, pjcontainer.PJContained):
         return f'<{self.__class__.__name__} ({self.__name__}) at {self._p_oid}>'
 
 
+
+@zope.interface.implementer(interfaces.IRevisionedImmutableManager)
 class ImmutableContainer(pjcontainer.AllItemsPJContainer):
 
     _pj_mapping_key = 'name'
@@ -130,16 +132,70 @@ class ImmutableContainer(pjcontainer.AllItemsPJContainer):
     def getCurrentRevision(self, obj):
         return self[obj.__name__]
 
-    def getRevisionHistory(self, obj):
+    def getRevision(self, name, version):
+        qry = self._combine_filters(
+            super()._pj_get_resolve_filter(),
+            sb.Field(self._pj_table, self._pj_mapping_key) == name,
+            sb.Field(self._pj_table, 'version') == version,
+        )
+        with self._pj_jar.getCursor() as cur:
+            cur.execute(
+                sb.Select(self._get_sb_fields(()), qry),
+                flush_hint=[self._pj_table])
+            row = cur.fetchone()
+        return self._load_one(
+            row[self._pj_id_column], row[self._pj_data_column], use_cache=False)
+
+    def getNumberOfRevisions(self, obj):
+        # 1. Setup the basic query.
         qry = self._combine_filters(
             super()._pj_get_resolve_filter(),
             sb.Field(self._pj_table, self._pj_mapping_key) == obj.__name__
         )
+        with self._pj_jar.getCursor() as cur:
+            cur.execute(sb.Select('COUNT(*)', qry), flush_hint=[self._pj_table])
+            return cur.fetchone()[0]
+
+    def getRevisionHistory(
+            self, obj, creator=None, comment=None,
+            startBefore=None, startAfter=None,
+            batchStart=0, batchSize=None, reversed=False):
+
+        # 1. Setup the basic query.
+        qry = self._combine_filters(
+            super()._pj_get_resolve_filter(),
+            sb.Field(self._pj_table, self._pj_mapping_key) == obj.__name__
+        )
+
+        # 2. Apply all additional filters.
+        if creator is not None:
+            creatorFld = sb.Field(self._pj_table, 'creator')
+            qry = self._combine_filters(qry, creatorFld == creator)
+        if comment is not None:
+            commentFld = sb.Field(self._pj_table, 'comment')
+            qry = self._combine_filters(qry, commentFld.contains(comment))
+        if startBefore is not None:
+            startOnFld = sb.Field(self._pj_table, 'startOn')
+            qry = self._combine_filters(qry, startOnFld < startBefore)
+        if startAfter is not None:
+            startOnFld = sb.Field(self._pj_table, 'startOn')
+            qry = self._combine_filters(qry, startOnFld > startAfter)
+
+        # 3. Setup ordering.
         orderBy = sb.Field(self._pj_table, 'version')
         fields = self._get_sb_fields(())
+
+        # 4. Apply batching.
+        batchEnd = None
+        if batchSize is not None:
+            batchEnd = batchStart + batchSize
+
+        # 5. Execute query.
         with self._pj_jar.getCursor() as cur:
             cur.execute(
-                sb.Select(fields, qry, orderBy=orderBy),
+                sb.Select(
+                    fields, qry, start=batchStart, end=batchEnd,
+                    orderBy=orderBy, reversed=reversed),
                 flush_hint=[self._pj_table])
             for row in cur:
                 obj = self._load_one(